moved file AbstractEmail/php into Abstract folder and updated tests
