+++ src/Abstracts/AbstractEmail.php
@@ -1,5 +1,5 @@
 <?php
-namespace Pokedex;
+namespace Pokedex\Abstracts;
 
 abstract class AbstractEmail
 {
