+++ test/src/ValueObjects/Email.php
@@ -1,7 +1,7 @@
 <?php
 use PHPUnit\Framework\TestCase;
 use Pokedex\ValueObjects\Email;
-require_once ('../../../src/AbstractEmail.php');
+require_once ('../../../src/Abstracts/AbstractEmail.php');
 require_once ('../../../src/ValueObjects/Email.php');
 require_once ('../../../src/Exceptions/EmailException.php');
